requirements.txt
# Build system
setuptools>=45
wheel

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0

# Validation & Serialization
pydantic[email]>=2.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# File Upload
python-multipart>=0.0.6

# Environment & Configuration
python-dotenv>=1.0.0

# Testing
requests>=2.31.0

# WebSocket
websockets>=12.0

# Optional Integrations (install as needed)
# twilio>=8.0.0  # For SMS
# boto3>=1.28.0  # For AWS S3
# openai>=1.0.0  # For OpenAI
# python-multipart>=0.0.6  # For file uploads